+++ cid-minting/tests/test_cid_minting_store/setup_cid_minter_db.sql
